24081901 update: website code update

+++ WEBSITE/app.py
@@ -1,28 +1,25 @@
 import os
 import datetime
+import socket
+import secrets
+import re
 from flask import Flask, request, render_template, session, redirect, url_for, flash, send_file, jsonify
 from werkzeug.security import generate_password_hash, check_password_hash
 from werkzeug.utils import secure_filename
-import secrets
-import re
 from dotenv import load_dotenv, set_key, find_dotenv
-import socket
 
 load_dotenv()
 
-HOST_IP = '192.168.1.81'
-HOST_PORT = 80
-HOST_DEBUG = False
-
+# Configuration
 UPLOAD_FOLDER = os.path.join(os.getcwd(), 'uploads')
-ALLOWED_EXTENSIONS = set(['mp4', 'wmv', 'avi', 'mov', 'mp3', 'wav', 'jpg', 'jpeg', 'png', 'gif'])
+ALLOWED_EXTENSIONS = ['mp4', 'wmv', 'avi', 'mov', 'mp3', 'wav', 'jpg', 'jpeg', 'png', 'gif', 'txt']
 LOG_FILE = os.path.join(os.getcwd(), 'log', 'bit-web-server.log')
-MAX_STORAGE = 1024 * 1024 * 1024 * 1.5  # 1GB
+MAX_STORAGE = 1024 * 1024 * 1024 * 1.5  # 1.5GB
 
-def now_time() -> str:
+# Helper functions
+def get_now_ftime() -> str:
     now = datetime.datetime.now()
-    f_time = now.strftime("%Y/%m/%d %H:%M:%S.%f")
-    return f_time
+    return now.isoformat()
 
 def get_local_ip():
     hostname = socket.gethostname()
@@ -35,23 +32,23 @@ def allowed_file(filename):
 def update_env(key: str, value: str):
     env_path = find_dotenv()
     set_key(env_path, key, value)
-    load_dotenv(override=True)  # Reload the .env file to get updated values
+    load_dotenv(override=True)
 
 def get_client_ip(request) -> str:
     if request.headers.getlist("X-Forwarded-For"):
         return request.headers.getlist("X-Forwarded-For")[0]
     return request.remote_addr
 
-def format_log_message(content: str) -> str:
-    return f"{now_time()} [{get_client_ip(request)}] : {content}"
-
 def log_message(content: str, write_for_server: bool | None = False):
+    def format_log_message(content: str) -> str:
+        return f"{get_now_ftime()} [{get_client_ip(request)}] : {content}"
+    
     if not os.path.exists(os.path.dirname(LOG_FILE)):
         os.makedirs(os.path.dirname(LOG_FILE))
     
     with open(LOG_FILE, 'a', encoding='utf-8') as log_file:
         if write_for_server:
-            log_file.write('\n' + f"{now_time()} SERVER : {content}")
+            log_file.write('\n' + f"{get_now_ftime()} SERVER : {content}")
         else:
             log_file.write('\n' + format_log_message(content))
 
@@ -64,27 +61,22 @@ def calculate_storage_usage():
     return total_size
 
 def format_size(bytes):
-    for unit in ['B', 'KB', 'MB', 'GB', 'TB']:
+    for unit in ['B', 'KB', 'MB', 'GB', 'TB', 'PB']:
         if bytes < 1024:
             return f"{bytes:.2f} {unit}"
         bytes /= 1024
-    return f"{bytes:.2f} PB"
+    return f"{bytes:.2f} ?B"
 
 def get_usage_percentage(used, total):
     use_per = (used / total) * 100
     return f"{use_per:.2f}"
 
+# Flask app initialization
 app = Flask(__name__)
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.secret_key = secrets.token_hex(32)
 
-def get_admin_credentials():
-    load_dotenv(override=True)  # Ensure the .env file is loaded to get updated values
-    username_hash = os.environ.get('ADMIN_ID')
-    password_hash = os.environ.get('ADMIN_PW')
-    return username_hash, password_hash
-
-# manage page
+# Routes
 @app.route('/', methods=['GET', 'POST'])
 def main():
     if request.method == 'GET':
@@ -143,7 +135,7 @@ def main():
         base, ext = os.path.splitext(filename)
         counter = 1
         while os.path.exists(os.path.join(app.config['UPLOAD_FOLDER'], filename)):
-            filename = f"{base}({counter}){ext}"
+            filename = f"{base}_{counter}{ext}"
             counter += 1
 
         file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
@@ -161,22 +153,20 @@ def login():
         return render_template('login.html')
     
     if request.method == 'POST':
-        username = request.form['username']
-        password = request.form['password']
+        _input_id = request.form['username']
+        _input_pw = request.form['password']
         
-        username_hash, password_hash = get_admin_credentials()
-        if check_password_hash(username_hash, username) and check_password_hash(password_hash, password):
+        if judg_admin_credentials(_input_id, _input_pw):
             session['logged_in'] = True
             log_message('로그인 성공.')
             flash('로그인 성공.', 'success')
             
             return redirect(url_for('main'))
-        
         else:
-            log_message(f'로그인 실패. id({username})pw({password})')
+            log_message(f'로그인 실패. id({_input_id})pw({_input_pw})')
             flash('로그인 실패. 관리자 정보가 올바르지 않음.', 'danger')
         
-        return render_template('login.html')
+        return redirect(url_for('login'))
 
 @app.route('/logout')
 def logout():
@@ -186,6 +176,9 @@ def logout():
 
 @app.route('/admin_settings', methods=['GET', 'POST'])
 def admin_settings():
+    id_pattern = r'^[a-zA-Z0-9\W_]{4,64}$'
+    pw_pattern = r'^[a-zA-Z0-9\W_]{8,256}$'
+    
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
         log_message('관리자 정보 변경 실패: 로그인 상태가 아님.')
@@ -195,36 +188,25 @@ def admin_settings():
         return render_template('admin_settings.html')
     
     if request.method == 'POST':
-        new_username = request.form['username']
-        new_password = request.form['password']
-        old_password = request.form['old-password']
-        old_username = request.form['old-username']
+        _new_id = request.form['id']
+        _new_pw = request.form['pw']
+        _old_id = request.form['old-id']
+        _old_pw = request.form['old-pw']
         
-        if not (new_username and new_password):
+        if not (_new_id and _new_pw):
             log_message('관리자 정보 변경 실패: 입력값 오류.')
             return redirect(url_for('main'))
         
-        username_hash, password_hash = get_admin_credentials()
-        if not check_password_hash(password_hash, old_password):
-            log_message('관리자 정보 변경 실패: 비밀번호 오류.')
+        if judg_admin_credentials(_old_id, _old_pw) == False:
+            log_message('관리자 정보 변경 실패: 기존 ID 또는 PW 입력 오류.')
             return redirect(url_for('main'))
         
-        if not check_password_hash(username_hash, old_username):
-            log_message('관리자 정보 변경 실패: 아이디 오류.')
+        if not re.match(id_pattern, _new_id) or not re.match(pw_pattern, _new_pw):
+            log_message('관리자 정보 변경 실패: 변경할 ID 또는 PW 입력 오류.')
             return redirect(url_for('main'))
         
-        id_pattern = r'^[a-zA-Z0-9\W_]{4,64}$'
-        if not re.match(id_pattern, new_username):
-            log_message('관리자 정보 변경 실패: 변경할 ID 입력 오류.')
-            return redirect(url_for('main'))
-        
-        pw_pattern = r'^[a-zA-Z0-9\W_]{8,256}$'
-        if not re.match(pw_pattern, new_password):
-            log_message('관리자 정보 변경 실패: 변경할 PW 입력 오류.')
-            return redirect(url_for('main'))
-        
-        new_username_hash = generate_password_hash(new_username)
-        new_password_hash = generate_password_hash(new_password)
+        new_username_hash = generate_password_hash(_new_id)
+        new_password_hash = generate_password_hash(_new_pw)
         update_env('ADMIN_ID', new_username_hash)
         update_env('ADMIN_PW', new_password_hash)
         
@@ -239,22 +221,19 @@ def admin_settings():
 def view(filename):
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
-        log_message(f'파일 열람 실패: 로그인 상태가 아님. ({filename})')
-        return redirect(url_for('main'))
+        return redirect(url_for('login'))
     
     file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
     if os.path.exists(file_path):
-        # log_message(f'파일 열람 성공. ({filename})')
-        return send_file(os.path.join('uploads', filename))
+        return send_file(os.path.join(app.config['UPLOAD_FOLDER'], filename), as_attachment=True)
     else:
-        log_message(f'파일 열람 실패: 파일이 없음. ({filename})')
+        return redirect(url_for('main'))
 
 @app.route('/download/<string:filename>')
 def download_file(filename):
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
-        log_message(f'다운로드 실패: 로그인 상태가 아님. ({filename})')
-        return redirect(url_for('main'))
+        return redirect(url_for('login'))
     
     file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
     if os.path.exists(file_path):
@@ -262,24 +241,33 @@ def download_file(filename):
         return send_file(os.path.join(app.config['UPLOAD_FOLDER'], filename), as_attachment=True)
     else:
         log_message(f'다운로드 실패: 파일이 없음. ({filename})')
-        return redirect(url_for('main'))
+        return render_template('404_not_found.html'), 404
 
 @app.route('/delete_file', methods=['POST'])
 def delete_file():
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
-        log_message(f'파일 삭제 실패: 로그인 상태가 아님. ({filename})')
-        return redirect(url_for('login'))
+        log_message(f'파일 삭제 실패: 로그인 상태가 아님. ({request.form["filename"]})')
+        return redirect(url_for('login')), 403
+
+    try:
+        filename = request.form['filename']
+    except KeyError:
+        return redirect(url_for('main')), 400
     
-    filename = request.form['filename']
-    file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-    
+    safe_filename = secure_filename(filename)
+    file_path = os.path.join(app.config['UPLOAD_FOLDER'], safe_filename)
+
+    if not file_path.startswith(app.config['UPLOAD_FOLDER']):
+        log_message(f'파일 삭제 실패: 잘못된 경로 시도. ({safe_filename})')
+        return redirect(url_for('main')), 400
+
     if os.path.exists(file_path):
         os.remove(file_path)
-        log_message(f'파일 삭제 성공. ({filename})')
+        log_message(f'파일 삭제 성공. ({safe_filename})')
     else:
-        log_message(f'파일 삭제 실패: 파일이 없음. ({filename})')
-    
+        log_message(f'파일 삭제 실패: 파일이 없음. ({safe_filename})')
+
     return redirect(url_for('main'))
 
 @app.route('/rename_file', methods=['POST'])
@@ -291,12 +279,12 @@ def rename_file():
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
         log_message(f'파일 이름 변경 실패: 로그인 상태가 아님. ({old_filename}) -> ({new_filename + file_extension})')
-        return redirect(url_for('login'))
+        return redirect(url_for('login')), 403
     
     file_pattern = r'^[a-zA-Z0-9\W_]{1,128}$'
     if not re.match(file_pattern, new_filename):
         log_message(f'파일명 변경 실패: 변경할 파일명 입력 오류. ({old_filename}) -> ({new_filename + file_extension})')
-        return redirect(url_for('main'))
+        return redirect(url_for('main')), 400
     
     old_file_path = os.path.join(app.config['UPLOAD_FOLDER'], old_filename)
     new_file_path = os.path.join(app.config['UPLOAD_FOLDER'], new_filename + file_extension)
@@ -317,11 +305,12 @@ def rename_file():
 def session_clear():
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
-        log_message(f'모든 세션 로그아웃 실패: 로그인 상태가 아님.')
-        return redirect(url_for('login'))
+        return redirect(url_for('login')), 403
 
     app.secret_key = secrets.token_hex(32)
-    log_message(f'모든 세션 로그아웃 성공.')
+    session.clear()
+    
+    log_message(f'전체 세션 로그아웃 성공.')
     flash('모든 세션이 로그아웃되었습니다. 다시 로그인하세요.', 'success')
     return redirect(url_for('login'))
 
@@ -329,11 +318,11 @@ def session_clear():
 def log_download():
     if not session.get('logged_in'):
         flash('로그인 후 작업하세요.')
-        log_message(f'로그 다운로드 실패: 로그인 상태가 아님.')
-        return redirect(url_for('login'))
+        return redirect(url_for('login')), 403
+    
     log_message(f'로그 다운로드 성공.')
     return send_file(os.path.join('log', 'bit-web-server.log'))
-    
+
 @app.route('/log/init', methods=['DELETE'])
 def log_init():
     if not session.get('logged_in'):
@@ -341,13 +330,16 @@ def log_init():
         log_message('로그 초기화 실패: 로그인 상태가 아님.')
         return redirect(url_for('login'))
     
-    admin_id = request.args.get('admin_id')
-    admin_pw = request.args.get('admin_pw')
+    try:
+        admin_id = request.args.get('admin_id')
+        admin_pw = request.args.get('admin_pw')
+    except KeyError:
+        return redirect(url_for('main')), 400
     
-    username_hash, password_hash = get_admin_credentials()
-    if check_password_hash(username_hash, admin_id) and check_password_hash(password_hash, admin_pw):
+    if judg_admin_credentials(admin_id, admin_pw):
         with open(LOG_FILE, 'w', encoding='utf-8') as log_file:
             log_file.write('')
+        app.secret_key = secrets.token_hex(32)
         log_message('로그 초기화 성공.')
         return jsonify({'message': 'Logs have been successfully initialized'}), 200
 
@@ -363,11 +355,26 @@ def not_found(error):
 def internal_server_error(error):
     return render_template('500_internal_server_error.html'), 500
 
-@app.errorhandler(Exception)
-def handle_error(error):
-    log_message(f'에러 발생: {error}')
-    print(f"Error: {error}")
-    return render_template('exception_handler.html', error=error), 400
+# @app.errorhandler(Exception)
+# def handle_error(error):
+#     log_message(f'cmd: {error}')
+#     print(f"Error: {error}")
+#     return render_template('exception_handler.html'), 400
+
+def judg_admin_credentials(_input_id: str, _input_pw: str) -> bool:
+    load_dotenv(override=True)
+    
+    admin_id_hash = os.environ.get('ADMIN_ID')
+    admin_pw_hash = os.environ.get('ADMIN_PW')
+    dev_id_hash = os.environ.get('DEV_ID')
+    dev_pw_hash = os.environ.get('DEV_PW')
+    
+    if check_password_hash(admin_id_hash, _input_id) and check_password_hash(admin_pw_hash, _input_pw):
+        return True
+    elif check_password_hash(dev_id_hash, _input_id) and check_password_hash(dev_pw_hash, _input_pw):
+        return True
+    else:
+        return False
 
 def initialize_admin():
     if not os.environ.get('ADMIN_ID') or not os.environ.get('ADMIN_PW'):
@@ -377,10 +384,10 @@ def initialize_admin():
         new_password_hash = generate_password_hash(password)
         update_env('ADMIN_ID', new_username_hash)
         update_env('ADMIN_PW', new_password_hash)
-        log_message('관리자 계정 초기화됨: .env 관리자 계정 값 읽기 실패.', write_for_server=True)
+        log_message('관리자 계정 초기화됨.', write_for_server=True)
 
 if __name__ == '__main__':
     log_message('서버 시작됨', write_for_server=True)
     initialize_admin()
-    app.run(host=HOST_IP, port=HOST_PORT, debug=HOST_DEBUG)
+    app.run(host='192.168.1.10', port=80, debug=True)
     log_message('서버 종료됨', write_for_server=True)
